Week3/AI_Agent_Tools_experiment/requirements.txt
python-dotenv
langchain
langchain-community
langchain-google-genai
google-search-results
